# _updateGlobalFundingGrowth() completed and twapInterval added.

+++ contracts/lib/PerpMath.sol
@@ -25,7 +25,7 @@ library PerpMath {
         return FullMath.mulDiv(priceX96, 1 ether, FixedPoint96.Q96);
     }
 
-    function formatX10_18ToX96(uint160 valueX10_18) internal pure returns (uint256) {
+    function formatX10_18ToX96(uint256 valueX10_18) internal pure returns (uint256) {
         return FullMath.mulDiv(valueX10_18, FixedPoint96.Q96, 1 ether);
     }
